+++ packages/media/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.0.0](https://github.com/maxst-fe/maxverse-web-sdk/compare/@maxverse/media-web-sdk@2.0.0...@maxverse/media-web-sdk@3.0.0) (2023-07-06)
+
+
+### Features
+
+* add features of screen sharing ([ca9bfd4](https://github.com/maxst-fe/maxverse-web-sdk/commit/ca9bfd4e0e2fb3f2a1f7a14beaf7720a9e7e259c)), closes [#7](https://github.com/maxst-fe/maxverse-web-sdk/issues/7)
+* add local track's  method of handle screen sharing ([9e98cf4](https://github.com/maxst-fe/maxverse-web-sdk/commit/9e98cf4ca2321df928ad79802068801729a1c607))
+* change types of participant event handlers ([3d0a516](https://github.com/maxst-fe/maxverse-web-sdk/commit/3d0a516d350800e1ecdbbdd57335b18aaae6201e)), closes [#7](https://github.com/maxst-fe/maxverse-web-sdk/issues/7)
+* **participant/eventContainer:** add conditional case for screen share ([f1d2b10](https://github.com/maxst-fe/maxverse-web-sdk/commit/f1d2b105935c233e429baaa973f7baaa3b86da65))
+* **participant/targetParticipant:** add a getter for isScreenShareEnabled in the TargetParticipant ([39cb40e](https://github.com/maxst-fe/maxverse-web-sdk/commit/39cb40e4bb205514454c03c14fb8abb003ecec3e)), closes [#7](https://github.com/maxst-fe/maxverse-web-sdk/issues/7)
+* **types/index.ts:** add types of participant event handler for screen sharing ([acea71b](https://github.com/maxst-fe/maxverse-web-sdk/commit/acea71b45d924e880c972b8112efca67c4da83bf))
+
+
+### BREAKING CHANGES
+
+* Change method related to control track
+
+- Change the following targetParticipant method
+ - attachTrackToElement
+ - detachTrackFromElement
+
+Please update your code to remove usages of these elements as they are no longer supported.
+Refer to the documentation(https://doc.maxverse.io/maxverse-sdk-media) for migration instructions.
+* Deprecated handlers related to types of OnLocalTrackPublished, OnTrackSubscribed
+
+- Change the following participant event handler:
+ - onLocalVideoTrackPublished => onLocalVideoUpdated
+ - onLocalAudioTrackPublished => onLocalAudioUpdated
+ - onVideoTrackSubscribed => onRemoteVideoUpdated
+ - onAudioTrackSubscribed => onRemoteAudioUpdated
+
+Please update your code to remove usages of these elements  as they are no longer supported.
+Refer to the documentation for migration instructions.
+
+
+
+
+
 # [2.0.0](https://github.com/maxst-fe/maxverse-web-sdk/compare/@maxverse/media-web-sdk@1.0.0...@maxverse/media-web-sdk@2.0.0) (2023-07-04)